feature: add resource generators to model transformer
