+++ packages/amplify-graphql-model-transformer/src/resources/rds-model-resource-generator.ts
@@ -1,6 +1,8 @@
 import * as cdk from '@aws-cdk/core';
 import { TransformerContextProvider } from '@aws-amplify/graphql-transformer-interfaces';
 import { ModelResourceGenerator } from './model-resource-generator';
+import {DatasourceType} from '@aws-amplify/graphql-transformer-core';
+import {DynamoDBModelVTLGenerator, ModelVTLGenerator, RDSModelVTLGenerator} from '../resolvers';
 
 /**
  * An implementation of ModelResourceGenerator responsible for generated CloudFormation resources
@@ -9,6 +11,11 @@ import { ModelResourceGenerator } from './model-resource-generator';
 export class RdsModelResourceGenerator extends ModelResourceGenerator {
   protected generatorType = 'DynamoModelResourceGenerator';
 
-  generateResources(scope: cdk.Construct, ctx: TransformerContextProvider): void {
+  generateResources(ctx: TransformerContextProvider): void {
+  }
+
+  // eslint-disable-next-line class-methods-use-this
+  getVTLGenerator(): ModelVTLGenerator {
+    return new RDSModelVTLGenerator();
   }
 }
