+++ packages/amplify-graphql-model-transformer/src/types.ts
@@ -1,0 +1,6 @@
+import { SyncConfig } from '@aws-amplify/graphql-transformer-core';
+
+export type ModelTransformerOptions = {
+  EnableDeletionProtection?: boolean;
+  SyncConfig?: SyncConfig;
+};
